requirements.txt
google-adk
google-generativeai
ddgs
feedparser
python-dotenv
